+++ windows_auth/settings.py
@@ -1,10 +1,19 @@
 from dataclasses import dataclass, field, asdict, MISSING, fields
-from typing import Dict, Any, Optional, Iterable, Union, Tuple
+from typing import Dict, Any, Optional, Iterable, Union, Tuple, List
 
 from django.core.exceptions import ImproperlyConfigured
 
 # domain name to use as a fallback setting for domain missing from WAUTH_DOMAINS
 DEFAULT_DOMAIN_SETTING = "__default__"
+
+
+def _get_group_list(value) -> List[str]:
+    if value is None:
+        return []
+    elif isinstance(value, str):
+        return [value]
+    else:
+        return value
 
 
 @dataclass()
@@ -39,7 +48,9 @@ class LDAPSettings:
     SUPERUSER_GROUPS: Optional[Union[str, Iterable[str]]] = "Domain Admins"
     STAFF_GROUPS: Optional[Union[str, Iterable[str]]] = "Administrators"
     ACTIVE_GROUPS: Optional[Union[str, Iterable[str]]] = None
-    GROUP_MAP: Dict[str, str] = field(default_factory=dict)
+    PROPAGATE_GROUPS: bool = True
+    GROUP_MAP: Dict[str, Union[str, Iterable[str]]] = field(default_factory=dict)
+    FLAG_MAP: Dict[str, Union[str, Iterable[str]]] = field(default_factory=dict)
 
     @classmethod
     def for_domain(cls, domain: str):
@@ -72,3 +83,30 @@ class LDAPSettings:
            for setting, value in merged_settings.items()
            if setting in cls_fields
         })
+
+    def get_superuser_groups(self):
+        return _get_group_list(self.SUPERUSER_GROUPS)
+
+    def get_staff_groups(self):
+        if self.PROPAGATE_GROUPS:
+            return list({*_get_group_list(self.STAFF_GROUPS), *self.get_superuser_groups()})
+        else:
+            return _get_group_list(self.STAFF_GROUPS)
+
+    def get_active_groups(self):
+        if self.PROPAGATE_GROUPS:
+            return list({*_get_group_list(self.ACTIVE_GROUPS), *self.get_staff_groups()})
+        else:
+            return _get_group_list(self.ACTIVE_GROUPS)
+
+    def get_flag_map(self):
+        return {
+            "is_superuser": self.get_superuser_groups(),
+            "is_staff": self.get_staff_groups(),
+            "is_active": self.get_active_groups(),
+            **{
+                k: _get_group_list(v)
+                for k, v in self.FLAG_MAP.items()
+            }
+        }
+